+++ src/Continy.php
@@ -535,11 +535,6 @@ class Continy implements Container
         return $output;
     }
 
-    public static function concatName(string $className, string $methodName): string
-    {
-        return $className . '::' . $methodName;
-    }
-
     public static function formatName(mixed $maybeCallable): string
     {
         if (is_array($maybeCallable) && 2 === count($maybeCallable)) {
@@ -553,6 +548,54 @@ class Continy implements Container
         }
 
         return (string)$maybeCallable;
+    }
+
+    /**
+     * @param string|array|callable $callback
+     *
+     * @return callable|null
+     * @throws ContinyException
+     * @throws ContinyNotFoundException
+     */
+    public function parseCallback(string|array|callable $callback): ?callable
+    {
+        if (is_callable($callback)) {
+            return $callback;
+        }
+
+        if (is_string($callback) && str_contains($callback, '@')) {
+            $split = explode('@', $callback, 2);
+        } else {
+            // array.
+            $split = $callback;
+        }
+
+        if (2 === count($split)) {
+            // 'foo@bar' style.
+            $cls    = $split[0];
+            $method = $split[1];
+
+            if (class_exists($cls) && method_exists($cls, $method)) {
+                if (is_callable([$cls, $method])) {
+                    // Static methods.
+                    return [$cls, $method];
+                }
+                // Common methods, the class needs to be instantiated,
+                // Or $cls may be an alias for the container.
+                $instance = $this->get($cls);
+                if (is_callable([$instance, $method])) {
+                    return [$instance, $method];
+                }
+            }
+        } elseif (1 === count($split)) {
+            // It may be a class name, a container alias.
+            $instance = $this->get($split[0]);
+            if (is_callable($instance)) {
+                return $instance;
+            }
+        }
+
+        return null;
     }
 
     public function getMain(): string