+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 1.2.0
+
+2025-07-14
+
+- Update `bojaghi/contract` to 1.3.
+- Remove `Continy::concatName` method.
+
 ## 1.1.5
 
 2025-07-14